pyproject.toml
[tool.poetry]
name = "es-semantic-search"
version = "1.0.0"
description = ""
authors = ["aa"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.13.1"
transformers = "^4.26.0"
pyvi = "^0.1.1"
elasticsearch = "^8.6.0"
pandas = "^1.5.3"
flask = "^2.2.2"
python-dotenv = "^0.21.1"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
